Optimization
The API allowing Lua to list out Kthura objects was waaay to slow.
That was because it was heavily relying on TList objects, which are when used for random access (which was needed for access by Lua) slowing everything down.
This commit contains an optimized version, which copies everything into an array, and can thus in radom access be much faster.

In [Star Story](http://github.com/phantasarproductions/star-story) the placement of enemies and random treasures can speed up greatly as a result.
[The Fairy Tale REVAMPED](http://github.com/phantasarproductions/TFTREVAMP) is not yet scripted out, so we'll never be able to compare the old situation with the new there, but it's very likely that game will benefit too in the treasure map builder (which I yet have to script).

And please ignore the .app bundles. Those are illegally compiled apps I forgot to remove before copying this crap to the repository, so I shall before I push this commit delete all that crap. If you are a Mac user, don't even try to run those .app bundles. None of them has a binary file, so they won't run. ;)

Jeroen